# docs/swagger.yml
swagger: '2.0'
info:
    title: User API
    description: Happy to code User API
    version: '1.0.0'
schemes:
    - http
basePath: /api/v1
produces:
    - application/json
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
paths:
    /users/login:
        post:
            summary: Login API
            description: |
                API for User Login
            parameters:
                - name: data
                  in: body
                  description: json body data for user login
                  required: true
                  schema:
                      required:
                          - username
                          - password
                      properties:
                          username:
                              type: string
                              description: User Name
                          password:
                              type: string
                              description: Password of User
            responses:
                200:
                    description: OK
                401:
                    description: Unauthorized
                400:
                    description: Bad Request
                406:
                    description: Not Acceptable
                500:
                    description: Server Internal error
            tags:
                - Users
    /users/create:
        post:
            summary: Create a new user account
            description: |
                Create a new user account
            parameters:
                - name: data
                  in: body
                  description: json body data to create new user
                  required: true
                  schema:
                      required:
                          - username
                          - email
                          - password
                          - firstname
                          - lastname
                      properties:
                          username:
                              type: string
                              description: User Name
                          password:
                              type: string
                              description: Password of User
                          email:
                              type: string
                              description: Email of User
                          firstname:
                              type: string
                              description: First Name
                          lastname:
                              type: string
                              description: Last Name
            responses:
                200:
                    description: Returns a access token
                401:
                    description: Unauthorized
                400:
                    description: Bad Request
                406:
                    description: Not Acceptable
                500:
                    description: Server Internal error
            tags:
                - Users

    /users/get/{id}:
        get:
            summary: Get a user
            description: |
                Get a user by Id
            security:
                - Bearer: []
            parameters:
                - name: id
                  in: path
                  description: User Id
                  required: true
                  type: string
                  format: string
            responses:
                200:
                    description: OK
                401:
                    description: Unauthorized
                400:
                    description: Bad Request
                406:
                    description: Not Acceptable
                500:
                    description: Server Internal error
            tags:
                - Users

    /users/list/{page}/{limit}:
        get:
            summary: Get list of user
            description: |
                Get list of user with page and limit
            security:
                - Bearer: []
            parameters:
                - name: page
                  in: path
                  description: Page Number
                  required: true
                  type: string
                  format: string
                - name: limit
                  in: path
                  description: Limit Number
                  required: true
                  type: string
                  format: string
            responses:
                200:
                    description: OK
                401:
                    description: Unauthorized
                400:
                    description: Bad Request
                406:
                    description: Not Acceptable
                500:
                    description: Server Internal error
            tags:
                - Users
    /users/update/{id}:
        put:
            summary: Update a user
            description: |
                Update a user with method put
            security:
                - Bearer: []
            parameters:
                - name: id
                  in: path
                  description: User Id
                  required: true
                  type: string
                  format: string
                - name: data
                  in: body
                  description: json body data to update user
                  required: true
                  schema:
                      required:
                          - username
                          - email
                          - password
                      properties:
                          username:
                              type: string
                              description: User Name
                          password:
                              type: string
                              description: Password of User
                          email:
                              type: string
                              description: Email of User
            responses:
                200:
                    description: OK
                401:
                    description: Unauthorized
                400:
                    description: Bad Request
                406:
                    description: Not Acceptable
                500:
                    description: Server Internal error
            tags:
              - Users

    /products/create:
        post:
            summary: Create a new product
            description: |
                Create a new product
            parameters:
                - name: data
                  in: body
                  description: json body data to create new product
                  required: true
                  schema:
                      required:
                          - name
                          - description
                      properties:
                          name:
                              type: string
                              description: Name
                          description:
                              type: string
                              description: Description of Product
                          amount:
                              type: number
                              description: Amount
                          in_stock:
                              type: boolean
                              description: In stock or not?
            responses:
                200:
                    description: Returns a access token
                401:
                    description: Unauthorized
                400:
                    description: Bad Request
                406:
                    description: Not Acceptable
                500:
                    description: Server Internal error
            tags:
                - Products
    /products/list/{page}/{limit}:
        get:
            summary: Get list of product
            description: |
                Get list of product with page and limit
            security:
                - Bearer: []
            parameters:
                - name: page
                  in: path
                  description: Page Number
                  required: true
                  type: string
                  format: string
                - name: limit
                  in: path
                  description: Limit Number
                  required: true
                  type: string
                  format: string
            responses:
                200:
                    description: OK
                401:
                    description: Unauthorized
                400:
                    description: Bad Request
                406:
                    description: Not Acceptable
                500:
                    description: Server Internal error
            tags:
                - Products
    /products/get/{id}:
        get:
            summary: Get a product
            description: |
                Get a product by Id
            security:
                - Bearer: []
            parameters:
                - name: id
                  in: path
                  description: Product Id
                  required: true
                  type: string
                  format: string
            responses:
                200:
                    description: OK
                401:
                    description: Unauthorized
                400:
                    description: Bad Request
                406:
                    description: Not Acceptable
                500:
                    description: Server Internal error
            tags:
                - Products